General clean-up

+++ ios/broccoli/Level.swift
@@ -1,18 +1,10 @@
-//
-//  Level.swift
-//  broccoli
-//
-//  Created by Nuno on 18/05/2017.
-//  Copyright © 2017 nunoalexandre. All rights reserved.
-//
-
 import Foundation
 
 class Level {
-    let value : Double
+    let value : Int
     
     init(_ value: String) {
-        self.value = Double(["one" : 1, "two" : 2,"three" : 3, "five" : 5, "eight" : 8, "twenty_one" : 21][value]!)
+        self.value = ["one" : 1, "two" : 2,"three" : 3, "five" : 5, "eight" : 8, "twenty_one" : 21][value]!
     }
-    
+
 }
